# services/contact/internal/delivery/http/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "contact service on clean architecture",
        "title": "slurm contact service on clean architecture",
        "contact": {
            "name": "API Support",
            "email": "mensheninao@gmail.com"
        },
        "license": {
            "name": "mensheninao"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/contacts": {
            "get": {
                "description": "Метод позволяет получить список контактов.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Получить список контактов.",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение при получении записей",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "name",
                        "description": "Сортировка по полю",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список контактов",
                        "schema": {
                            "$ref": "#/definitions/contact.ListContact"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Метод позволяет создать контакт.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Метод позволяет создать контакт.",
                "parameters": [
                    {
                        "description": "Данные по контакту",
                        "name": "contact",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contact.ShortContact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "201": {
                        "description": "Структура контакта",
                        "schema": {
                            "$ref": "#/definitions/contact.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/contacts/{id}": {
            "get": {
                "description": "Метод позволяет получить контакт по мдентификатору контакта.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Получить контакт.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор контакта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура контакта",
                        "schema": {
                            "$ref": "#/definitions/contact.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Метод позволяет обновить данные контакта.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Метод позволяет обновить данные контакта.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор контакта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные по контакту",
                        "name": "contact",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contact.ShortContact"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Структура контакта",
                        "schema": {
                            "$ref": "#/definitions/contact.ContactResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Метод позволяет удалить контакт.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Метод позволяет удалить контакт.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор контакта",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "404 Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "contact.ContactResponse": {
            "type": "object",
            "required": [
                "createdAt",
                "id",
                "modifiedAt",
                "phoneNumber"
            ],
            "properties": {
                "age": {
                    "description": "Возраст",
                    "type": "integer",
                    "default": 0,
                    "maximum": 200,
                    "minimum": 0,
                    "example": 42
                },
                "createdAt": {
                    "description": "Дата создания контакта",
                    "type": "string"
                },
                "email": {
                    "description": "Электронная почта",
                    "type": "string",
                    "format": "email",
                    "maxLength": 250,
                    "example": "example@gmail.com"
                },
                "gender": {
                    "description": "Пол",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "id": {
                    "description": "Идетификатор записи",
                    "type": "string",
                    "format": "uuid",
                    "example": "00000000-0000-0000-0000-000000000000"
                },
                "modifiedAt": {
                    "description": "Дата последнего изменения контакта",
                    "type": "string"
                },
                "name": {
                    "description": "Имя клиента",
                    "type": "string",
                    "maxLength": 50,
                    "example": "Иван"
                },
                "patronymic": {
                    "description": "Отчество клиента",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Иванович"
                },
                "phoneNumber": {
                    "description": "Мобильный телефон",
                    "type": "string",
                    "maxLength": 50,
                    "example": "78002002020"
                },
                "surname": {
                    "description": "Фамилия клиента",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Иванов"
                }
            }
        },
        "contact.ListContact": {
            "type": "object",
            "properties": {
                "limit": {
                    "description": "Количество записей",
                    "type": "integer",
                    "default": 10,
                    "minimum": 0,
                    "example": 10
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/contact.ContactResponse"
                    }
                },
                "offset": {
                    "description": "Смещение при получении записей",
                    "type": "integer",
                    "default": 0,
                    "minimum": 0,
                    "example": 20
                },
                "total": {
                    "description": "Всего",
                    "type": "integer",
                    "default": 0,
                    "minimum": 0,
                    "example": 10
                }
            }
        },
        "contact.ShortContact": {
            "type": "object",
            "required": [
                "phoneNumber"
            ],
            "properties": {
                "age": {
                    "description": "Возраст",
                    "type": "integer",
                    "default": 0,
                    "maximum": 200,
                    "minimum": 0,
                    "example": 42
                },
                "email": {
                    "description": "Электронная почта",
                    "type": "string",
                    "format": "email",
                    "maxLength": 250,
                    "example": "example@gmail.com"
                },
                "gender": {
                    "description": "Пол",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "name": {
                    "description": "Имя клиента",
                    "type": "string",
                    "maxLength": 50,
                    "example": "Иван"
                },
                "patronymic": {
                    "description": "Отчество клиента",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Иванович"
                },
                "phoneNumber": {
                    "description": "Мобильный телефон",
                    "type": "string",
                    "maxLength": 50,
                    "example": "78002002020"
                },
                "surname": {
                    "description": "Фамилия клиента",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Иванов"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "info": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}